Official v1.2.4 Alpha Version *Redone threading system that uses 1 thread instead of many. *Changed the updater to latest, which should give correct version if the admin is keeping up to date with CraftBukkit. *Added a few new API features (as always). *Added a few more separate classes for threads.
EXPLAINED THREADING:
There is now 1 main thread. This main thread gathers all other thread
classes from an arraylist, and runs them accoring to their interval and
count, and if they are told to run or not. This prevents the use of many
threads, and puts all threading into 1 main thread that runs every tick.